Update equivalences to use new data structures
